requirements.txt
# external requirements
Sphinx
python-dotenv>=0.5.1
numpy>=1.14
scikit-image>=0.13
imageio>=2.3
matplotlib>=3.0
GDAL>=2.2
Werkzeug>=0.14.1
Flask>=1.0.2
requests>=2.18.4
PyYAML>=3.12
Keras>=2.2
